+++ SYCL/ESIMD/regression/Inputs/complex-lib-sycl.cpp
@@ -6,7 +6,7 @@ sycl::event iota(size_t n, sycl::buffer<int, 1> &buf, sycl::queue &Q) {
     auto K = [=](sycl::id<1> id) {
       int *y = acc_y.get_pointer();
       size_t i = id.get(0);
-      y[i] = i;
+      y[i] = static_cast<int>(i);
     };
     H.parallel_for(n, K);
   };
